+++ src/Pattern/StdPatterns/DatePattern.php
@@ -55,7 +55,7 @@ class DatePattern extends AbstractPattern
             return (new \DateTime())->setTimestamp($data)->format(static::DATE_FORMAT);
         }
 
-        if (is_string($data) && preg_match('#^(' . $this->getRegex() . ')$#', $data)) {
+        if (is_string($data) && $this->match($data)) {
             $sData = (string) $data;
             if ($this->checkDate($sData)) {
                 return $sData;
@@ -77,7 +77,8 @@ class DatePattern extends AbstractPattern
             }
         }
 
-        if (is_string($data) && preg_match('#^(' . Patterns::REGEX_INT . ')$#', $data)) {
+        $d = Patterns::DELIMITER;
+        if (is_string($data) && preg_match($d . '^(' . Patterns::REGEX_INT . ')$' . $d, $data)) {
             return (int) $data;
         }